requirements.txt
absl-py>=1.4.0
astor>=0.8.1
certifi>=2023.7.22
chardet>=5.2.0
cycler>=0.12.1
future>=0.18.3
gast>=0.5.4
grpcio>=1.62.1
gym==0.26.2
h5py>=3.10.0
idna>=3.6
joblib>=1.3.2
keras==2.14.0
keras-preprocessing>=1.1.2
kiwisolver>=1.4.5
markdown>=3.5.2
matplotlib>=3.7.4
mock>=5.1.0
numpy>=1.23.5
pandas>=2.2.2
pillow>=10.3.0
protobuf>=4.25.3
pyglet==1.5.27
pyparsing>=3.1.2
python-dateutil>=2.9.0
pytz>=2024.1
requests>=2.31.0
scipy>=1.13.0
six>=1.16.0
tensorboard>=2.14.0
tensorflow==2.14.0
tensorflow-estimator==2.14.0
termcolor>=2.4.0
urllib3>=2.2.1
werkzeug>=3.0.2